feat: Add functionality to write current date to notes.txt file
This commit adds the functionality to write the current date to the notes.txt file. It initializes the project with the package.json file and creates the app.js file to generate the current date using the `Date` object. It also imports the `fs` module to write the date to the notes.txt file.